refactor and move docs
