+++ std/pubkey_test.go
@@ -1,19 +1,21 @@
-package codec_test
+package std_test
 
 import (
 	"testing"
 
-	"github.com/cosmos/cosmos-sdk/crypto/codec"
-	"github.com/cosmos/cosmos-sdk/crypto/types/multisig"
+	"github.com/cosmos/cosmos-sdk/std"
+
 	"github.com/stretchr/testify/require"
 	"github.com/tendermint/tendermint/crypto"
 	"github.com/tendermint/tendermint/crypto/ed25519"
 	"github.com/tendermint/tendermint/crypto/secp256k1"
 	"github.com/tendermint/tendermint/crypto/sr25519"
+
+	"github.com/cosmos/cosmos-sdk/crypto/types/multisig"
 )
 
 func roundTripTest(t *testing.T, pubKey crypto.PubKey) {
-	cdc := codec.DefaultPublicKeyCodec{}
+	cdc := std.StdPublicKeyCodec{}
 
 	pubKeyEnc, err := cdc.Encode(pubKey)
 	require.NoError(t, err)
@@ -22,7 +24,7 @@ func roundTripTest(t *testing.T, pubKey crypto.PubKey) {
 	require.Equal(t, pubKey, pubKeyDec)
 }
 
-func TestDefaultPublicKeyCodec(t *testing.T) {
+func TestStdPublicKeyCodec(t *testing.T) {
 	pubKeySecp256k1 := secp256k1.GenPrivKey().PubKey()
 	roundTripTest(t, pubKeySecp256k1)
